KTS_APP_TRLRGRAM/requirements.txt
aio-pika==9.0.5
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aioredis==2.0.1
aiormq==6.7.4
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
botocore==1.29.76
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.12
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pamqp==3.2.1
pydantic==1.10.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
six==1.16.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.8.2
sqlalchemy==2.0.9